Create basic SQLAlchemy models and import prod data

+++ dev-env/backend/patch-last-migration.sh
@@ -6,19 +6,17 @@ set -o pipefail
 cd "$(dirname "${BASH_SOURCE[0]}")"
 
 
-docker compose exec backend-shell ./manage.py migrate textbooks zero
-docker compose exec --workdir /app/backend/gabby backend-shell alembic downgrade head-1
+rev_id_arg=
+if [ -e ../../backend/gabby/migrations/versions/*_dev.py ]
+then
+  rev_id_arg="--rev-id $(find ../../backend/gabby/migrations/versions/*_dev.py | sed -E 's#../../backend/gabby/migrations/versions/(.*)_dev\.py#\1#')"
+  docker compose exec --workdir /app/backend/gabby backend-shell alembic downgrade head-1
+  rm -f ../../backend/gabby/migrations/versions/*_dev.py
+else
+  docker compose exec --workdir /app/backend/gabby backend-shell alembic upgrade head
+fi
 
-rm -f ../../backend/textbooks/migrations/0007_*.py
-rm -f ../../backend/gabby/migrations/versions/*_dev.py
+docker compose exec --workdir /app/backend/gabby backend-shell alembic revision --autogenerate $rev_id_arg -m dev
 
-docker compose exec backend-shell ./manage.py makemigrations
-docker compose exec --workdir /app/backend/gabby backend-shell alembic revision --autogenerate -m dev
-
-docker compose exec backend-shell ./manage.py migrate
 docker compose exec --workdir /app/backend/gabby backend-shell alembic upgrade head
-
-docker compose exec backend-shell ./manage.py loaddata test-exercises more-test-exercises
-
-docker compose exec backend-shell ./manage.py graph_models textbooks --disable-sort-fields --rankdir BT --output textbooks/models.png
 docker compose exec backend-shell python -m gabby graph-models
